+++ TBS/TBS/Player.cs
@@ -19,8 +19,8 @@ public class Player : Combatant
 
     public List<Item> equippedItems = new List<Item>(4);
 
-    public float inventoryWeight = 0;
-    public float inventorySpeedModifier = 0;
+    public int inventoryWeight = 0;
+    public int inventorySpeedModifier = 0;
 
     public List<Attack> ownedAttacks = new List<Attack>();
